user registration added
a very simple user registration page was added (email/password). This
will need to be expanded with any other essential user information.
We’ll also need a page for updating information, a temporary table to
hold registrations until email addresses are validated, and add in
means for linking in google/facebook/etc accounts while creating or
logging in to accounts (and to recover passwords)