updated code to work with schema changes
updated code to work with schema changes
moved session_start() to dbconfig.php. Honestly, it probably belongs
somewhere else. We might want a dbconfig for storing the login info and
another config that is always loaded first. Either that, or we should
have a single config that is always loaded and find another way to
store the login info for the DB (this is probably what we should do)