band-aid fix for bug in larval
This is actually a bug in laravel — it should be accessing the config
parameter for the proper table name instead of assuming the default
(‘users’). We should fix this manually later, but as long as our table
is named ‘USER’, this won’t be a problem. Here is a link describing the
problem:

http://laravel.io/forum/02-16-2015-laravel-ignores-user-table