updated USER table to be compatible with eloquent
Eloquent, a means of accessing a DB built in to Laravel, is an option
for authentication. In our config, we chose DIY instead. However, it
looks as though the authentication system in laravel still insists on
some aspects of eloquent. It expects that there are two timestamp
fields: created_at and updated_at, that exist within the user table. I
believe this is another bug (though it’s possible we have something
misconfigured — I did set the auth configuration to “database” and not
“eloquent”), but I’m currently working around it by adding in those
fields. I’m not sure it’s such a bad idea to have them anyway.